Add files via upload
I started the logistic regression model following our class notes on the subject. When I get to the line for YhatTestProb, I get an error. I have included the error in the code file. I'll keep looking for solutions but if anyone else can see something wrong, please let me know. I'm wondering if it is related somehow to the fact that our supervisor is quantitative and in the notes and homework we used a qualitative supervisor. You'll notice that I had to coerce Ytrain from a dataframe to numeric to get the outLogistic model to accept it.